--- Aplikasi penyewaan fasilitas/PenyewaanFasilitas/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=e8f83e4d
-build.xml.script.CRC32=ed69492c
-build.xml.stylesheet.CRC32=f85dc8f2@1.112.0.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=e8f83e4d
-nbproject/build-impl.xml.script.CRC32=4ddc4a58
-nbproject/build-impl.xml.stylesheet.CRC32=12e0a6c2@1.112.0.48
