--- Aplikasi penyewaan fasilitas/PenyewaanFasilitas/nbproject/private/private.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/Damai/Aplikasi%20penyewaan%20fasilitas/PenyewaanFasilitas/src/penyewaanfasilitas/Register.java</file>
-            <file>file:/C:/Users/Damai/Aplikasi%20penyewaan%20fasilitas/PenyewaanFasilitas/src/penyewaanfasilitas/Login.java</file>
-            <file>file:/C:/Users/Damai/Aplikasi%20penyewaan%20fasilitas/PenyewaanFasilitas/src/penyewaanfasilitas/PenyewaanFasilitas.java</file>
-            <file>file:/C:/Users/Damai/Aplikasi%20penyewaan%20fasilitas/PenyewaanFasilitas/src/penyewaanfasilitas/Home.java</file>
-        </group>
-    </open-files>
-</project-private>
