+++ nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=C:\\Users\\ilham\\AppData\\Roaming\\NetBeans\\22\\build.properties
