+++ src/penyewaanfasilitas/PenyewaanFasilitas.java
@@ -1,0 +1,25 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Main.java to edit this template
+ */
+package penyewaanfasilitas;
+
+/**
+ *
+ * @author Damai
+ */
+public class PenyewaanFasilitas {
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String[] args) {
+        // TODO code application logic here
+        Login lg = new Login();
+        lg.setVisible(true);
+        lg.pack();
+        lg.setLocationRelativeTo(null);
+        lg.setDefaultCloseOperation(Login.EXIT_ON_CLOSE);
+    }
+    
+}
